+++ lib/dyph3.rb
@@ -1,4 +1,4 @@
 require "dyph3/version"
 require "dyph3/differ"
-require "dyph3/heckel_diff"
-require "dyph3/resig_diff"
+require "dyph3/two_way_differs/heckel_diff"
+require "dyph3/two_way_differs/resig_diff"
