+++ lib/dyph3/examples/diff3.pl
