+++ lib/dyph3/examples/moin_differ.rb
