+++ lib/dyph3/examples/penn_differ.py
